emails: Move organisation logo attachment to EmailAplus
All mail classes inheriting from EmailAplus need to attach the organisation
logo - so move the code there. Fixes certain emails lacking any logo.
