+++ apps/newsletters/emails.py
@@ -1,5 +1,3 @@
-from email.mime.image import MIMEImage
-
 from django.apps import apps
 from django.conf import settings
 from django.contrib import auth
@@ -35,18 +33,6 @@ class NewsletterEmail(ReportToAdminEmailMixin, Email):
             .filter(get_newsletters=True)\
             .filter(is_active=True)\
             .distinct()
-
-    def get_attachments(self):
-        attachments = super().get_attachments()
-
-        organisation = self.kwargs['organisation']
-        if organisation and organisation.logo:
-            f = open(organisation.logo.path, 'rb')
-            logo = MIMEImage(f.read())
-            logo.add_header('Content-ID', '<{}>'.format('organisation_logo'))
-            attachments += [logo]
-
-        return attachments
 
 
 class NewsletterEmailAll(NewsletterEmail):
